+++ lgsm/modules/command_update.sh
@@ -38,6 +38,6 @@ else
 fi
 
 # remove update lockfile
-rm -f "${lockdir}/${selfname}-update.lock"
+rm -f "${lockdir:?}/${selfname}-update.lock"
 
 core_exit.sh
